refactor: cleanup and handler improvements
- Remove Docker configuration files
- Simplify service configuration
- Update account, project, and project integration handlers
- Clean up types and constants
- Update request/response models

+++ crates/nvisy-postgres/src/client/pg_database.rs
@@ -51,7 +51,7 @@ impl PgClient {
             AsyncDieselConnectionManager::new_with_config(&config.database_url, manager_config);
 
         let pool = Pool::builder(manager)
-            .max_size(config.pool.max_size)
+            .max_size(config.pool.max_size as usize)
             .wait_timeout(Some(config.pool.connection_timeout))
             .create_timeout(Some(config.pool.connection_timeout))
             .recycle_timeout(config.pool.idle_timeout)
